feat: define baseline keep-alive versus close behavior
- Introduced ConnectionBehavior enum to manage connection states.
- Implemented connection_behavior_for method to determine connection behavior based on request headers.
- Updated response serialization to conditionally set the Connection header based on the determined behavior.
- Modified success_response method to accept connection behavior as a parameter.
- Enhanced request head failure responses to include appropriate connection behavior.

closes: #32
