+++ gems/vajra/ext/vajra/response/response_writer.hpp
@@ -20,7 +20,7 @@ namespace Vajra
     {
     public:
       bool send(int client_fd, const Response &response) const;
-      Response success_response() const;
+      Response success_response(ConnectionBehavior connection_behavior = ConnectionBehavior::close) const;
       Response request_head_failure_response(Vajra::request::HeadFailureKind kind) const;
       void log_request_head_error(const Vajra::request::HeadError &error) const;
